Test suite and some other nice things
git-svn-id: https://guust.tuxes.nl/svn/bas/studie/0910-afstuderen/Implementations/c++@419 840f488b-753d-4fb4-b916-1b28b8a18e6e

+++ WAHStackTC/src/PerformanceTimer.h
@@ -19,10 +19,10 @@ public:
 	virtual ~PerformanceTimer();
 
 	static PerformanceTimer start();
-	static long diffTimeMilliSecs(const timeval& time1, const timeval& time2);
+	static double diffTimeMilliSecs(const timeval& time1, const timeval& time2);
 	static long diffTimeMicroSecs(const timeval& time1, const timeval& time2);
-	long reset();
-	long currRunTime();
+	double reset();
+	double currRunTime();
 	long currRunTimeMicro();
 };
